backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0

# Email parsing
mail-parser==3.15.0

# File type detection
python-magic==0.4.27

# Office macro analysis
oletools==0.60.1

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.2.1

# YARA pattern matching
yara-python==4.5.1

# PDF analysis
pdfminer.six==20231228
pypdf==4.2.0

# Hashing / crypto

# Archive inspection
rarfile==4.1

# PE file analysis
pefile==2023.2.7

# Async utilities
anyio==4.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
